adds useSearch custom hook, but not everything is smooth.
